move registry and config
