+++ config/kube/watcher_test.go
@@ -1,4 +1,4 @@
-package config
+package main
 
 import (
 	"context"
